Fixed bugs and added multiple.

+++ kunnon-kansalainen/model.py
@@ -1,7 +1,7 @@
 from google.appengine.ext import db
 class KunnonKansalainen(db.Model):
   action=db.StringProperty()
-  time=db.IntegerProperty()
+  time=db.FloatProperty()
   unit=db.StringProperty()
   freq=db.StringProperty()
   total_time=db.IntegerProperty()
